AI_Image_Generator/requirements.txt
torch>=2.0.0
transformers>=4.25.1
diffusers>=0.14.0
scipy
streamlit
Pillow